OPAL-193. Decoupling the Aria connection from the Diagnosis module.
The Diagnosis module is now independent of Aria. If
Aria is not active, the diagnostics from Aria are
masked. If no diagnostics list is empty, an error msg
is displayed saying no add or modifications can be
done. If Aria should be active but there is a DB
problem, an error msg will be displayed.