OPAL-194. Decoupling the Aria connection from the Lab Results module.
The Lab Results module is now independent of Aria. If
Aria is not active, the lab results from Aria are
masked. If the lab results list is empty, an error msg
is displayed saying no add or modifications can be
done. If Aria should be active but there is a DB
problem, an error msg will be displayed.