OPAL-195. Decoupling the Aria connection from the Filter class.
The Filter class is now independent of Aria. If
Aria is not active, only the filters from OpalDB are
shown in Post, Educational Material and Publication
Tools section (add and edit). If Aria should be active
but there is a DB problem, an error msg will be
displayed in the appropriate section.