Fixed the internal error and we can edit Question
I fixed the internal error by changing "!==" to "!=" in each comparison.
In the "if" statement, the type changed from the Old question to the new question.
Since "!==" return true if it's not identical, that means not equal or not the same type,
we got in the if statement and return $answer to false.

By changing to "!=", we ignore the type change, and we do not get in the if function.


(cherry picked from commit 999b64ff053a5da0e0268eb93a7e6d8b014b2000)

Co-authored-by: Benoit <151671747+Benoite142@users.noreply.github.com>